Adds Preference Pane along with default settings
Need to test more. Working but need to make adjustments to prevent overwriting existing settings file. Also need to add logic if there is an informational alert (20) that it is unable to check for updates, the next run will need to check if the issue has been fixed and a full run can be done.